requirements.txt
pytesseract
Pillow
keyboard
win10toast
pyperclip